+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 Change Log
 
+v3.0.0
+---
+* **Breaking change:** `ignoreRequireImports` option renamed to `ignoreImports`
+* **Breaking change:** `rotateStringArray` option renamed to `stringArrayRotate`
+* **Breaking change:** `shuffleStringArray` option renamed to `stringArrayShuffle`
+* `ignoreImports` now ignores `dynamic` imports in addition to `require` imports
+
 v2.19.1
 ---
 * Updated `@javascript-obfuscator/escodegen` with fixed parenthesis of `default export` IIFE